refactor:update src/api to append message_controller for  path
